Add resource.ExpressionFailedError
This class is meant to be consumed by the UI layers to provide meaningful
error messages to the operator. The expression attribute can be used to
obtain the text of the expression that failed as well as the resource name
that is used by that expression. The resource name can be used to lookup
the (resource) job that produces such values.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>