Add the session module.
This is probably the most important complicate part of the entire
checkbox/plainbox rewrite. The class defined there SessionState
represents a testing "session". The session is both a low-level
state wrappe and high-level state presenter for the UI.

In practical terms it knows about all jobs, resources and results.
Internally it computes which jobs are runnable and which are blocked
by readiness inhibitors. There are several kinds of inhibitors for
common problems such as job having a failed dependency or not having
a resource. Read that part of the code for details as it is heavily
documented. The session also encapsulates the dependency solver and any problems
it might report.

Notably, the session does not run any jobs. This vastly simplifies
the code and makes it much easier to test. All tests are based on
setting up a session object with some initial list of known jobs,
setting up a subset of that jobs as 'desired' (the ones that we want to
run) to observe how runnability of each job changes.

The session also updates the state it manages when presented with
job results. Each job result is analyzed and for resource and local
jobs, the output is also parsed.

The code is mostly fully tested. There are some edge cases that need
more work around the problems reported by the dependency solver in the
update_desired_job_list() call but causal testing showed that no obvious
catastrophic bugs exist in that area.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>