runner: use CommandOutputWriter instead of Redirect
Previously we've used Redirect to create verbatim log files on disk.
This was somewhat ugly for clean-up/termination point of view as stuff
might fail and open files would be left around until recycled by the
python garbage collector.

The CommandOutputWriter is a little smarter, opening the file only when
the command actually starts executing

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>