checkbox: allow loading jobs directly via checkbox instance
Previously loading jobs was historically done by the load(somewhere)
function. This is really odd as it was only really loading checkbox
jobs and had this information hardcoded. With the wake of decoupling
the core from checkbox we need to couple checkbox pieces together.

In practical terms a CheckBox instance can now discover and load all
jobs by itself.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>