IO logs kepts are now kept on disk
First it will save memory but the crucial advantage is the ability
to restore them on resume.
Only the filename of the json dump is saved in the session file.
The io_log get method will open the corresponding json file, b64decode
the data msg to create a new "in memory" io log object.