checkbox: refactor how checkbox is located
This refactors large parts of CheckBox class. It is a side effect of
discovering that checkbox was packaged differently (different filesystem
layout) in precise than quantal or raring. After implementing a "quick
fix" for that I've decided to refactor and simplify the code around the
concept of 'directory helper' that knows where certain directories are.

Now there are three "layouts" for directories that are checked in oreder:

1) Source layout, called 'src': this is basically the checkbox tree
as it exists in the checkout of lp:checkbox

2) Deb layout for 12.10 and 13.04, called 'deb2': this is the current
layout for packaged checkbox where the script directory is in
/usr/lib/checkbox/bin

3) Deb layout for 12.04 called 'deb1': this is the older layout where
the scripts directory was in /usr/share/checkbox/scripts/

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>