depmgr: iterate over dependencies in deterministic order
The DependencySolver._visit() job iterates over each dependency as
computed by the DependencySolver._get_dependency_set() method. This
returns a set so the ordering is unspecified.

It makes sense to make this part deterministic so that we always run
jobs in the same order. It also fixes random "problems" for integration
tests that check the verbatim copy of the output.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>