job: Consider any environment variable from the config object, not only those declared in the job's environ.
We still give preference to the real environment; the config can't override an existing environment variable.

This is needed for some jobs (e.g. environment resource) that need access to not only its environment variables,
but also those set explicitly in the config. This also allows any jobs that aren't running as root to access those
explicitly-set environment variables.