secure: trusted-launcher: Added the code to run checkbox job commands as root
The trusted launcher is the minimal code needed to be able to ask the root
password only once.
This launcher has to be started using pkexec so that the related policy file
works as expected. The goal is to match a given job hash to a system-wide
available job and execute its command.

For jobs generated from local jobs the trusted launcher in started
with --via meaning that we have to eval a local job to find a hash match

Once a match is found, the job command is executed using os.execve.
The environment mapping replaces the previous implementation where we had to
call the env(1) command. Key/value pairs are optional positional arguments.
But only variables defined in the job environ property are allowed to avoid
compromising the root environment.
Needed modifications like adding CHECKBOX_SHARE and new paths to scripts are
managed by the trusted launcher.