plainbox:session:state: add session metadata support
This patch adds a new class SessionMetadata which is being implicitly
used inside SessionState. The class offers three bits of non-critical
information that helps to classify the runtime state of a session.

Using session 'flags' applications can add arbitrary boolean states
to any session. Two flags are provided by the core but any additional
flags may be used.

The flags provided here are FLAG_SUBMITTED, to indicate that results
of this session have been submitted somewhere and FLAG_INCOMPLETE, to
indicate that this session is still in progress.

Using session 'title' applications that want to use multiple sessions
can name sessions in a way that is meaningful to the application. This
can offer users some aid in selecting the session to operate on.

Using session 'running_job_name' applications can indicate their intent
to execute a particular job before actually attempting to do so. In the
case of a critical failure applications can recover this information and
use it to alter application flow, for instance, by offering to skip or
mark the affected job as a failure

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>