plainbox:session:storage: add SessionStorageRepository and SessionStorage
The two classes added in this patch add a way to manage and enumerate
SessionStorage objects (which represent any data a SessionState may
need to put in a filesystem) in a particular location (abstracted
as SessionStorageRpository)

SessionStorage class has two basic methods for saving and loading
checkpoints (save_checkpoint, load_checkpoint) that can basically
write a bunch of data to a single file atomically. There are two
implementations of those, one for python3.2 and another for python3.3
or more recent. The latter version uses *at functions that finally
got exported to the python standard library. Using *at functions
(such as openat(2), unlinkat(2)) we have better guarantees against
filesystem races (as we can reliably open descriptors to directories
and perform any directory relative operations using the relative name
and the descriptor).

SessionStorage can also keep track of the last session created in a
particular repository. This part is a bit ugly (I don't really like it)
but it's a compromise that allows single-session applications to work
without major changes

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>