plainbox: split JobResult into MemoryJobResult and DiskJobResult
This large patch splits JobResult into two classes. The MemoryJobResult
class keeps the IO log in memory. It is suitable for unit tests
and results that have no IO at all (skips, dummys, etc).
The DiskJobResult class keeps everything in a file on disk (but does not
handle any writing, that must be done externally) using a gzipped
stream of IOLogRecord tuples, as managed by the two existing reader/writer
classes.

The whole tree is updated to use {Disk,Memory}JobResult instead of the
old JobResult class11. Most of the tests now use MemoryJobResult so
that they don't have to fiddle with all the scratch directories and
cleanup.

Some larger changes were applied to the JobRunner class. It no longer
creates the .json file at the end of the command execution. Instead
it just uses the .record.gz file that was already there (from earlier
patches) to construct a DiskJobResult object and pass it back.

The serialization code, as implemented in plainbox.impl.session.state
got some slight improvement so that it does not have to do linear
look-ups of class names.

The serialization format changes a little to differentiate the two
classes (the constants  mapping those classes have been changed)
and it is no longer possible to load older results this way.
Backwards compatibility is coming in the upcoming, versioned format.
Doing it now would be possible but needlessly complicated IMHO

Lastly to work around python bug http://bugs.python.org/issue10791
I had to implement a subclass of GzipFile that implements read1()
so that we could read text from gzip files using TextIOWrapper()

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>