plainbox:dbus:service: drop _get_property_iface_map()
This patch removes the useless duplicate (mostyl) code that I wrote
before realizing how _dbus_class_table works. In short, instead
of building a property-speicifc dict<interface,dict<prop_name, prop>>
we're just reusing the somewhat different _dbus_class_table.
This a lot of the code without dropping any feature.

There are some logging changes (improvements) and implementations of
Get/GetAll/Set were changed to match everything else. There is some
non-obvious and not-that-perfectly-tested code that is somewhat tricky
when python name of the property is not the same as the DBus name
of the property.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>