plainbox:dbus:service: combine Object and ObjectManager
This patch changes how ObjectManager interface is implemented.
In practice sub-classing all of the vanilla DBus binding classes is
a bit problematic, due to the existence of the meta-class and
_dbus_class_table. This makes writing a mix-in class that provides
only particular functionality (such as the ObjectManager interface)
rather difficult. Instead of building a range of base classes that
mix-and-match all of the standard interfaces that we want to support
we just offer one basic class, Object, that implements them all
but each one has to be enabled via an opt-in flag.

In practice the :meth:`Object.should_expose_interface()` now decides
if the Introspect() method mentions a particular interface or not.
By default only the Introspectable interface is enabled. All the other
interfaces can be enabled by declaring class or instance attribute
HIDDEN_INTERFACES with appropriate data.

All of the methods have been re-arranged and put in front of headers
that clearly document what kind of a method/attribute is being defined.
There are three basic cases: overridden things from vanilla DBus bindings,
new functions but standardized with DBus spec and lastly new functions
or attributes custom to this implementation.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>