plainbox:commands: add PlainBoxToolBase class that implements basic cli
This class is copied from box.py, with slight changes to main() and
usage of @abstractmethod on three key differentiation methods,
get_config_cls(), get_exec_name() and add_subcommands(). This was badly
needed so that subclasses of (the old copy of this class) didn't have
to un-plainboxize it. This was especially visible in cloudbox which
unlike checkbox-ng, looks less like checkbox than current plainbox did.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>