plainbox:service: rewrite the DBus 'service' module
This patch changes how the DBus service handles life-cycle of various
objects it creates and how it reacts to dynamic changes (generation of
additional jobs, substitution of results -- both of which happen during
the resume operation).

Instead of "patching" holes in the coverage of DBus wrappers for native
(python, non DBus) objects after each operation that can potentially
cause such operation to be required, the resume code registers event
listeners on key signals and performs the necessary actions as those
events are fired.

One important change is that generated jobs are now being "owned" by the
session wrapper object in which they became generated. In general, when
(eventually) we will support having multiple sessions at the same time,
each session may generate identical jobs and we cannot let those clash.
In addition, we cannot really add jobs to any of the providers as it
would make little sense.

As mentioned earlier the results and state objects are now created
dynamically when new jobs are added. While state object are never
removed, results can be both created and removed. This can happen, for
example, when resuming a session or when re-running a job.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>