plainbox:session:state: add SessionState.trim_job_list()
This new method can be used to remove unused jobs from the session.
All of the state associated with the removed jobs is also removed
(state, result and resources). The function ensures that none of the
removed jobs are on the run list. This is done mainly to prevent
complexities that would arise if we'd have to re-generate desired job
list which can be done separately be the application.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>