plainbox:exporter:xml: map OUTCOME_UNDECIDED to OUTCOME_SKIP
This patch prevents a crash when manual jobs that were executed in
non-interactive mode (those now return OUTCOME_UNDECIDED) would fail
to export to something that the XML consumers can expect.

It also adds a couple extra missing mappings in the xml exporter, and
a test to verify that all possible plainbox outcomes are properly
mapped in the exporter.