plainbox:service: track SessionState.on_job_removed
This patch allows SessionStateWrapper to keep track of the
new on_job_removed signal and remove jobs (and all associated stuff) from
DBus as this happens.

This feature, so far, is only needed when something calls Resume()
and the list of jobs is trimmed to contain only those that were previously
saved.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>