plainbox:rfc822: replace Origin.filename with Origin.source
This patch changes how the Origin class keeps track of the "source" of
the text that is being referred to by the remaining two instance
attributes. Previously the "filename" attribute would be a path to a
file which contains some RFC822 data. With this patch instead of
tracking explicit filenames we track opaque source objects. With the
existing three source classes we can now represent "unknown", "file" and
"python source" and distinguish that at runtime.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>