plainbox:rfc822: have source propagate from {load,gen}_rfc822_records
Both load_rfc822_records and gen_rfc822_records can now be supplied with
an explicit source object, that identifies the origin of the data in the
text stream.

The source can be None, where the old logic of inferring it from the
stream name is retained.

Unlike what was there before, the resulting origin is always non-None.
In the case where neither explicit source nor stream.name could be used,
origin is created with UnknownFileText source).

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>