plainbox:{job,session:state}: make JobDefinition.via a computed property
This patch changes how the via property gets computed. Previously we
would store _via as an instancve attribute (defaulting to None) and set
it in JobDefinition.create_child_job_from_record(), so that places that
construct "child" jobs could use that method to create "children".

Now, since JobDefinition.origin is a bit more generic than before, and
can store a "job" type of origin, we can compute via without storing
anything more.

To do this, the create_child_job_from_record() method now really ensures
that the origin.source of the record defining the new child job is a
JobOutputTextSource and that it actually traces back to the right job.
This required some small changes to existing tests that simply assumed
that was brute-force copied over.

One unexpected change (which I don't particularly like) is how
__category__-style jobs now override the origin (not mearly set the via
attribute) of existing jobs that were read from regular files. I guess
this is required until we finally dispose of them and store category as
a job attribute.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>