plainbox:session:legacy: use session_file in previous_session_file
This patch changes how the legacy session API is implements the
previous_session_file() method. Previously it would return the path of
the 'last-session' directory. The actual value didn't matter as
applications used this value just to see if it was None or not.

The old implementation could leat to 'phantom sessions' causing a crash
on resume where the session was indeed created (and might still be
active but the legacy api supports at most one session anyway) but was
never checkpointed (persistent_save() was never called). In that case
the directory would exist but the state file would not and resume would
crash.

This patch uses the newly introduced 'SessionStorage.session_file'
(which points at the file) property instead of the current
'SessionStorage.location' (which points at the directory) and only
returns that value if the file with that name exists.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>