plainbox:dbus:service: fix InterfacesAdded to report all interfaces
Previously the InterfacesAdded signal would only carry information
about interfaces that have properties. This patch makes it contain
information about all the interfaces.

I believe that this is in accordance with the DBus specification:

"""The InterfacesAdded signal is emitted when either a new object is added
or when an existing object gains one or more interfaces. The
InterfacesRemoved signal is emitted whenever an object is removed or it
loses one or more interfaces. The second parameter of the InterfacesAdded
signal contains a dict with the interfaces and properties (if any) that
have been added to the given object path. Similarly, the second parameter
of the InterfacesRemoved signal contains an array of the interfaces that
were removed. Note that changes on properties on existing interfaces are
not reported using this interface - an application should also monitor
the existing PropertiesChanged signal on each object."""

This patch fixes https://bugs.launchpad.net/checkbox/+bug/1236377

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>