plainbox:runner: Non automated jobs return OUTCOME_UNDECIDED
To let the client set the outcome, non automated jobs use this special
value to trigger the corresponding interaction.

The manual and user-interact plugins now return a MemoryJobResult
containing {'outcome': IJobResult.OUTCOME_UNDECIDED}).

The user-verify and user-interact-verify only alter the outcome of the
prerun command result, this way we preserve all logs and associated state.