plainbox:plugins: add FsPlugInCollection
This new plug in collection class looks for files with a specific
extension in the provided search path. The intent of this class it to
allow application code to define some plugin definition standard and
default search locations and allow anyone to provide plugins this way.
This is simplar to how dbus service files or desktop files are
discovered and used.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>