plainbox:providers:checkbox: drop the ChechBox{Deb,Auto}Provider
The Deb provider is now fully replaced by filesystem-based providers so
it does not need to exist as a separate class. The Auto provider was a
hack to select either the Src or the Deb provider but since the Deb
provider is now a collection of smaller providers than this feature
cannot be implemented using a simple this-or-that choice. The Src
provider will now be a special mode (similar to stub) where we *only*
use one provider and discard the rest.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>