plainbox: move config to secure package
This patch moves plainbox.impl.config to plainbox.impl.secure.config as
we intend to use the config module (and a few more) in the secure part
of the codebase. There are a few automated replacements across the tree
to make this work but there is no new code being added.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>