plainbox:job: drop JobDefinition.modify_execution_envirionment()
This patch drops the now-unused method form the JobDefinition class.
The intended goal is to drop all of the "complex" code away from
JobDefinition so that it can be fully moved and merged with BaseJob from
the plainbox.impl.secure.job module. This will finally allow to use
multiple providers inside the checkbox trusted launcher.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>