plainbox:runner: use execution controllers to actually run commands
This patch drops most of the hacky and long set of methods from
JobRunner, including _get_script_env(), _get_command_trusted(),
_get_command_src(), and bits of _prepare_io_handling() and _run_extcmd()
and replaces all of that with a selection of the best execution
controller for the particular job and a single method call.

Currently there is no way to specify execution controllers but that's
just because I didn't want to sprinkle API changes all over the tree.  A
patch will follow, most likely, with a way to specify the list of
execution controllers.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

drop tests for runner