checkbox-ng:commands:server: New subcommand
To run the Server Certification whitelist
It's basically a copy of the run sub command but without the ability to
select the output file/format and without any transport support.

Results are displayed to stdout saved as submission.xml for certification
purpose and also saved as HTML and XLSX (if xlsxwriter is installed) for
later use.