move certification transport to checkbox-ng
This patch moves the CertificationTransport class (which is totally
specific to Canonical and CheckBox to the checkbox-ng sub-project). As
before, unit tests and documentation are adjusted. This also _almost*
drops the requests dependency from plainbox and moves it to checkbox-ng
(but not quite, there is still some code in run.py that needs it). The
transport entry point is also moved to the checkbox-ng package.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>