move sru-specific variables to checkbox-ng
PlainBox needed several configuration variables when the sru command was
added. Now that it is moved to checkbox-ng it makes sense to move the
associated configuration variables as well.

Because of the way Config inheritance works this change is backwards
compatible as checkbox-ng will load sru-specific variables from both
plainbox and checkbox-ng configuration files.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>