checkbox-ng: mock-away requirements for readthedocs.org
This patch removes the need to install checkbox-ng dependencies for
setup.py when invoked by readthedocs.org. All the externally imported
modules are mocked in conf.py, by injecting a Mock() object into
sys.modules.

This operation has no negative effects for the documentation as we are
not relying on any real properties of those dependencies.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>