plainbox:ctrl: fix how nest_dir is inserted into PATH
This patch corrects one of the many faces of the 'path bug' (executables
not being in PATH in certain situations).

One aspect of that bug is how nest_dir (directory, that has symlinks to
executables we need in PATH) was being injected into PATH. Previously it
was being done in execute_job(). This fails for execution controllers
that pass environment using a trick in get_execution_command(), such as
pkexec and plainbox-trusted-launcher execution controllers. It fails
because those controllers need to pass "differential" environment inside
the command itself and the change occurs outside of
get_execution_environment().

The fix is straightforward, to move all PATH manipulations to happen
inside get_execution_environment() and adjust all tests that needed
adjusting.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>