plainbox:secure:providers: add Provider1.load_all_jobs()
This method acts just like get_builtin_job() that it replaces with two
important differences. First, it never raises RFC822SyntaxError, instead
it collects all such exceptions into a list. This allows it to keep
loading jobs from subsequent files even if the first file has a syntax
problem early on. Second is that the full list of problems is returned
to the caller.

This allows it to be more resilient in face of user errors and makes it
a valuable method to have for tools that are aimed at provider developers.

The old get_builtin_jobs() is retained for compatibility but has been
re-implemented on top of load_all_jobs().

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>