checkbox-ng:service: Job command results are now sent with AskForOutcome
A new signal called ShowInteractiveUI is now used instead of AskForOutcome
to just open the manual interaction UI element. Because there's no
suggested outcome in that case, it's up to the client to decide the default
outcome (usually IJobResult.OUTCOME_SKIP)

The AskForOutcome signal gets a new parameter - suggested_outcome - to carry
the result of the job command result.

Finally, both signals have been moved to the Session Wrapper class and are
now using the SESSION_IFACE.