plainbox:secure:config: add Config.write()
This patch adds a two methods to the Config class, culminating with the
Config.write() that writes a representation suitable to be read back by
Config.read(). The intermediate Config.get_parser_obj() method can be
used in use cases not covered by write() that still need access to the
same data.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>