plainbox:secure:config: Config._read_commit validate defaults
The _read_commit() method was called after either a string or a file
was parsed for variable and section definitions. Previously missing
variables (the ones we know about but we haven't found in the parsed text)
would be silently ignored. In particular, their validators would not
be checked.

This could cause later bugs where the caller assumes validators prevent
incorrect values from slipping into other parts of the program.

Now if a variable has no explicit definition, we use the default value
(which might be the Unset object) and "assign" that instead.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>