plainbox:clitools: work around argparse changes in 3.3.4
This patch adds a work-around for build failures spotted on Ubuntu
14.04 on the 28th of January 2014. We believe that those failures were
caused by the recent upload of python 3.3.4 which included updates to
the standard library argparse module. Those changes caused slight change
in the contents of the command usage help message.

For consistency with Ubuntu 12.04 which is our officially supported
target we're bundling a small fragment of argparse from python 3.3.1,
specifically the HelpFormatter._format_usage() method, along with
appropriate explanation and possible deprecation strategy.

The bundled fragment was written by:
        Steven J. Bethard <steven.bethard@gmail.com>.
It is covered by the PYTHON SOFTWARE FOUNDATION LICENSE VERSION 2.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>