plainbox:secure:qualifiers: add select_jobs()
This function implements the so-called whitelist-ordering where
a sequence of jobs is selected from an available set of jobs according
to a sequence of qualifiers. Fine structure of those qualifiers is
interrogated to ensure that jobs being selected by earlier fragments
of those qualifiers appear early on the resulting list of selections.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>