checkbox-ng:setup.py: import concurrent.futures
To avoid to display an error in atexit._run_exitfuncs after running tests.

Rationale:
Some tests import ThreadPoolExecutor from concurrent.futures.
That leads the concurrent.futures.thread module to be imported, and with it
an exit function to be registered:

atexit.register(_python_exit)

The problem seems to be that concurrent.futures which is loaded in by
unittest is then unloaded before the _exit_function gets called, and it's
function definitions by the way.