plainbox:job: New JobTreeNode class
JobTreeNode class is used to store a tree structure. A tree consists of a
collection of JobTreeNode instances connected in a hierarchical way where
nodes are used as categories, jobs belonging to a category are listed in the
node leaves.

The create_tree constructor can use a legacy mode using local jobs to build
the tree or an experimental job property called categories where a job can
belong to several categories.