plainbox:job: add JobDefinition.summary
The summary field is a short, human readable, name of the job. Unlike
job.name or the upcoming job.id, it can be translated. It is meant to be
displayed in user interfaces almost everywhere where job.name was was
previously used.

For a transitional period, job.summary is defaults to job.name. This
ensures that all jobs have a summary and that it is not automatically
changing its value.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>