plainbox:provider_manager: add 'manage.py i18n'
This patch adds the 'manage.py i18n' sub-command of the provider
management script. It allows provider authors to update, merge and build
translation catalogs for their jobs.

This command uses intltool-update internally (which pulls in a
dependency on perl, an gnu gettext). In the future we may transition to
python-babel and drop all those dependencies.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>