plainbox:secure:config: don't validate default values
Each Variable instance used to validate the default value but that
validation was inconsistent, it would be silently skipped if the default
value was Unset (which is the implicit default value).

After the generalization of validators and the introduction
of @understands_Unset we have way too many false positives that use
the NotUnsetValidator together with an Unset default to express
the requirement that some variable is defined.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>