plainbox:secure:providers: switch default gettext_domain to be Unset
This patch corrects the default value for
Provider1Definition.gettext_domain to Unset. This was not the case
before because the validation system was broken (with regards to default
values *and* unset used together). This is now fixed so we can simply
default to no domain.

The single place that uses Provider1Definition.gettext_domain now uses
'or None' to get None if Unset shows up.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>