plainbox:secure:providers: add Provider1.*_dir
This large patch discards all logic from Provider1 class that used to
compute various properties (such as jobs directory, whitelists
directory, etc). All of those properties are now explicit attributes of
Provider1 objects.

The logic is now entirely moved to Provider1PlugIn, which looks at a
Provider1Definition to either use explicit directory paths or generate
the missing data based on the same logic as before (relative to provider
location)

In addition to the logic move, some new directories are added:

The new Provider1.locale_dir property adds a way to associate locale
directory with each provider.  This directory can be fed to
bindtextdomain(). Using this together allows plainbox to load
translation catalogs for any provider without any hard-coded
assumptions.

The new Provider1.data_dir property exposes the existing 'data/'
directory that is common to many providers. It used to be related to the
CHECKBOX_SHARE directory and now CHECKBOX_SHARE is defined in terms of
data_dir (as the parent directory, whatever that is).

This allows providers without any clear location (such as the one with
translations, additional executables, program data, etc) to be defined
without ambiguities.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>