plainbox:secure:providers: add Provider1.from_definition()
This patch adds a convenience method encapsulating a lot of logic for
constructing Provider1 objects out of definitions (either created in
memory or loaded from the filesystem).

Two types of provider definitions can be loaded, one with the 'location'
field and one without it. The 'location' field is used as an anchor for
all the remaining directories (all are optional). If absent, all of the
directories need to be spelled out explicitly.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>