plainbox:secure:providers: make Provider1PlugIn smarter
This patch makes Provider1PlugIn contain all the logic of how to
instantiate a Provider1 from Provider1Definition. This includes
validating all of the _dir variables and making use (or not) of the
location attribute (which is kind of base directory for a provider that
keeps everything in one parent directory, e.g., for development).

Existing unit tests were simplified and refocused on testing the logic
unique to Provider1PlugIn. Some tests were moved to Provider1Definition.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>