plainbox:secure:providers: add Provider1Definition.{implicit,effective}_*
This patch adds a number of pairs of properties such as implicit_jobs_dir
and effective_jobs_dir that encapsulate the logic of using explicit values
inside the definition object or, if they are Unset, using the implicit
values computed from not-Unset location field but only if they would be
valid values themselves.

The exception is gettext_domain that has no implicit_value.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>