plainbox:secure:providers: bind per-provider gettext domain
This patch makes each constructed Provider1 object bind the gettext
domain it is using to the appropriate locale_dir, if both are defined
and true in a boolean context. This allows providers to see data in
under-development or otherwise custom locations as long as it is
accounted for in the provider definition file.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>