plainbox:ctrl: make execution controllers namespace-aware
This patch changes how the symlink nest is constructed. Previously
all of the providers would be consulted, making the effective symlink
nest a combination of all the existing providers. This would make
it possible for two provider authors to inadvertently clash with other
providers without any way of avoiding that. Now only providers that
from the same namespace as the namespace of the job that needs to be
executed are considered. This limits possible clashes to the scope
of the entity owning the namespace.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>