plainbox:commands:checkbox: add partial namespace support to run -w
This patch makes various commands that support `plainbox run` command
line interface support namespaces when handling -w | --whitelist.

Instead of naively loading the whitelist from a file, if that file
belongs to a known provider the whitelist is obtained through the
provider. This ensures that the implicit provider namespace is used to
qualify patterns and things 'just work' as before.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>