plainbox:provider_manager: fix unix layout provider generator
During provider-ization (introduction of normal manage.py) of our
current providers we found that in the UNIX layout the .provider file
entry associated with the each of the component directories (jobs_dir,
whitelists_dir, etc) was created unconditionally. This was problematic
for providers that didn't contain the full set of directories.

This patch makes the existence of each entry dependent upon the
existence of the corresponding directory in the source layout of the
provider.

A test definition was adjusted since the corresponding test was not
creating the locale directory.

Fixes: https://bugs.launchpad.net/checkbox/+bug/1297155

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>