plainbox:abc: add ISessionStateTransport interface
This interface describes all plainbox 'transports' which allow the user
to send the stream of exported session data after a sequence of tests.

This interface differs from the old de-facto interface it replaced by
moving additional, session-specific parameters to send(). It also
standardizes on passing both the configuration object and the session
state objects, along with the data stream to send, so that specialized
transports may access additional information.

Lastly the interface now specifies exceptions that may be raised as well
as the return value that is expected. This should allow anyone to use
arbitrary transport in an uniform way.

Fixes: https://bugs.launchpad.net/plainbox/+bug/1301967

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>