plainbox: LaunchpadTransport support in High Level API
Account for the needs of the LaunchpadTransport in the High Level API.
We need to pass the session state to send and make sure we account for the
possibility that the dictionary returned might also contain the 'state'
key instead of 'url'. Return a bad response message if neither of these
keys are present.

Signed-off-by: Brendan Donegan <brendan.donegan@canonical.com>