checkbox-ng:dbus: SelectJobs sees all session jobs
This patch changes how ServiceWrapper.SelectJobs() operates. Previously
it would look at all providers and select among any of the jobs defined
there. The problem is that that would never consider generated jobs
(children of local jobs that are not acting as a pure category
assignment).

This patch fixes that problem in a slightly hacky way. Since
SelectJobs() is defined on the service object (ideally it would have
been defined on the session object) and the whole DBus API cannot
support more than one session without breaking the API we can cheat and
peek at the full list of jobs as seen through the first (only) existing
session that the service object knows about. If the session does not
exist yet we just do what we did before, match among all the provider
jobs.

Fixes: https://bugs.launchpad.net/checkbox-ng/+bug/1309128

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>