plainbox:session: refactor _restore_SessionState_metadata()
This patch changes the API of
SessionResumeHelper{1,2,3}_restore_SessionState_metadata() so that
the first argument is a metadata object rather than a session object.

In addition, each of the 1-2-3 format versions implementations of that
method are moved to separate mix-in class MetaDataHelper{1,2,3}MixIn.
The existing SessionResumeHelper{1,2,3} clasess are all made to inherit
from the mix in.

This effectively allows the mix-in classes to be used by another
piece of code that wants to resume the meta-data object alone, without
having to resume the whole session.

Existing tests are adjusted to pass 'session.metadata' instead
of 'session', to compensate for the change.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>