Extract all per-commit tests to self-contained scripts.
Each top-level component (and each component under providers/) can
supply a test script named "container-tests-*", to be run to test the
container.

Scripts must adhere to the following contract:

Tests are found in each component's requirements/ dir and are named
*container-tests-* Tests scripts are expected to:
 - Have a name describing what they do (e.g.
   container-tests-plainbox-unit-testing)
 - Be run from the *component's* top-level directory. This is a
   convenience to avoid a boilerplate "cd .." on every test script. So
   for 'plainbox' we do the equivalent of
   $ cd $BLAH/plainbox
   $ ./requirements/container-tests-plainbox
 - Exit 0 for success, other codes for failure
 - Write logs/debugging data to stdout and stderr.
 - Scripts can be prefixed with a number (001-container-tests-blah) to
   force a specific running order, this works within each component
   only.

All existing tests were converted to this format. Note we're not using
them yet, it's preparation to support multiple container technologies
for testing. Tests for providers that didn't have them were also added,
as the script for providers is the same for all of them. Also a test for
the checkbox-support component was added.