checkbox-ng:cli: retain job ordering
This patch changes how selection from the tree menu is interpreted for
actual job selection in the session. Previously we would take the value
of the SelectableJobTreeNode.selected property. That would be okay but
the whitelist ordering of the original arguments to that tree is lost.

Now we take the same ``selected`` property and only use it as a mask to
select or reject jobs from the original list.

Fixes: https://bugs.launchpad.net/checkbox-ng/+bug/1308106

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>