plainbox:unit: add base unit module
Units are a new concept that will be the base of jobs and templates (and it
the future, test plans). This patch adds a simple unit class based on
parts of the JobDefinition class.

Each unit is associated with a provider and a RFC822 record (indirectly,
it carries an origin field though). This maps well to our existing jobs
and should match whitelists and templates without any issues.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>