plainbox:session:state: make SessionState() accept and understand units
This patch changes the initializer of SessionState to consume a list
of units instead of a list of jobs. This allows the session to store
templates (and in the future, test plans). The existing job_list property
is left unchanged, as is the instance attribute _job_list. It is computed
from unit_list by filtering out everything that is not a job definition.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>