plainbox:commands:run: add Action and _pick_action_cmd()
This patch adds a way for the run command to issue complex interactive
queries in an uniform way. By passing a list of actions (and an optional
prompt) to _pick_action_cmd() the user is presented with a list of
translated action labels, untranslated label accelerators (single-key
shortcuts) and the calling code gets an untranslated command string.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>