plainbox:result: change i18n-aware rendering of job outcome
The previous textual outcome was based on very short keywords like
none, pass, skip, etc which only really work for English with its very
simple sentence structure. In other languages (verified on Polish) it
was insufficient and impossible to correctly translate the outcome text
as a part of a sentence.

This patch makes two changes: first of all, the strings are now much
longer, giving the value a sense of a sentence and second of all, they
are contextualized and translated with pgettext, to allow translators
to localize that particular text precisely.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>