providers:checkbox: use existing WWAN connection for tests.
Previously, it was mandatory to set environment variables for WWAN
testing, so the script could automatically create the modem connection
and activate it. Unfortunately the best way to discover those settings
is to create a connection manually and then copy them from there. This
is OK for bulk testing, but for testing a single system, it creates
duplication of work.

These changes will skip the automatic connection creation and
destruction if a wwan connection already exists, so users can configure
their connection manually and leverage that for testing.

- If the environment variables are set, drop any existing connections,
  create the test connection before testing, then destroy it after
  testing.
- If environment variables are not set and an active connection is
  found, test using that one.
- If environment are not set and there is no active connection, fail the
  test.