plainbox:unit: implement Unit.__hash__()
This moves just a litte bit more code from JobDefintion to Unit. This time
the __hash__ method is implemented as the hash of the checksum (exactly
the same way as in JobDefinition right now).

I think all units should share implementation of this and several other
methods like __eq__ / __ne__ (coming up next)

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>