plainbox:resource: differentiate resource alias from resource id
This patch changes ResourceExpression to take a new optional argument
'imports' (expecting the return value of JobDefiniton.get_imported_jobs())
and to differentiate the identifier of the resource job an expression
depends on (resource_id) from the alias it has in the expression itself
(resorce_alias).

In addition, the imports argument is now searched for matching
aliases. If a matching alias is found then the resource identifier is
obtained from the import statement. This allows job definitions to
depend on resources from arbitrary namespace.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>