plainbox:unit: add unit metadata to the base Unit class
This patch adds a Meta class to the base Unit class. Inside the Meta
class there is a template_constraints field which aids the template unit
in instantiating units. Here each field of the base unit class (id and
unit) is given expected parametric data. A field can be either constant
('const'), when it is expected to be constant among all
template-instantiated units or variable ('vary'), where each
instantiated unit is expected to have a parametric value.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>