plainbox:unit: make the base Unit class parametrizable
This patch makes the base Unit class able to be customized with additional
parameter data. This is an optional supplementary dictionary with
arbitrary additional parameters.

There are additional APIs for working with parameteric units.
The ``is_parametric`` property and the ``parameters`` property offer
simple access to parameter data.

The ``get_accessed_parameters()`` method computes a mapping that describes
which parameters are used by each field defined in the unit.

The low-level field access methodS ``get_record_value()``
and ``get_raw_record_value()`` now respect and understand parameters and
will silently use parameters when required.

The ``validate()`` method is now checking parametric units for parameter
constraints. Those constraints are defined per-field and express
additional constraints such as ``vary`` or ``const`` that may be imposed
on any field.

The ``checksum`` property is now computed in a way that takes parameters
into account so that two identical units with different parameters
will get different checksums. Checksums of non-parametric units are
left unchanged.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>