plainbox:ctrl: instantiate template definitions after running resource jobs
This patch makes CheckBoxSessionStateController instantiate all the unit
templates that depend on a resource after that resource runs. This works
but there are two quirks about it.

First, we should not need to iterate over all the units like that (some
additional API to get a list of templates that depend on a resource
would be better).

Second it is going to misbehave when a resource is re-ran and produces a
different set of values. This needs some extra design decisions as it is
not immediately obvious how it should behave in that case. We might
remove all of the instantiated jobs and re-instantiate them with the new
set, we could reject this operation entirely or do something else.

It is not clear what the proper outcome should be though.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>