plainbox:provider_manager: display id if possible
The ./manage.py validate script currently displays the str() of the unit that
has some problems but this field is defined to be the "user readable" string.
This causes some pain when working with job definitions or templates that could
display the ID field instead. Now that we can use UnitWithId let's add a
hackish special-case for UnitWithId and JobDefinition (which is not an
UnitWithId yet) so that the id field is used instead.

Signed-off-by: Zygmunt Krynicki <zkrynicki@gmail.com>