plainbox:unit: copy all the "legacy" code to new _legacy module
This patch takes legacy validation code paths from *each* existing unit
and copies it to the new _legacy module. The module contains a
collection of classes ${name}ValidatorLegacyAPI and ${name}LegacyAPI
that contain legacy code for the validator and unit classes
respectively.

All of the validators were unified to be normal classes with normal
methods and not a random collection of classes with normal and static
methods. Some test code was adjusted not to talk to the validator
directly and instead call unit.validate()

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>