plainbox:unit:validators: fix compute_value_map()
This patch fixes an issue in the compute_value_map() helper function that would
previously crash with AttributeError if it attempted to probe an unit for a
property that that particular unit does not support.

This got evident when the file unit started to show up but obviously didn't
have the 'id' field. Now such units are simply ignored (since they don't have a
value for the field it makes sense). Note that this is distinct from having a
None value which is supported and works (and did work) correctly.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>