plainbox:{ctrl,runner,abc}: refactor execution controller API
This patch makes execution controller a little bit more stateless by moving
the ``session_dir`` argument away from the initializer and into the
IExecutionController.execute_job() method. This allows us to work towards
simplifications in the runner (which should be almost equally stateless).

Curently the job runner API is the same but it is now much easier to
delegate all of the responsibility into a SessionDeviceContext which is the
central point of associating a device to a provider.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>