plainbox:session:state: add SessionDeviceContext.execution_controller_list
This patch begins the move of responsibility to device awareness from various
places inside plainbox and into the device-specific code. Since we don't yet
have device classes all of that code is currently exposed by the
SessionDeviceContext facade.

The new property, execution_controller_list, allows applications to obtain a
list of execution controllers applicable to a given context. This will allow
the job runner to have an even smaller coupling to session internals and should
be capable of working with just a per-invocation context argument.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>