Cherry-pick revno 3316 from lp:checkbox.
plainbox:session:legacy: remove no-op operation from _commit_clean()

This patch removes an effectively state-effect-free call to
SessionManager.create_with_unit_list(). Since this call doesn't affect
the _manager instance (it is a class method so it cannot) the only side
effect was that it would create a new session (not even a legacy one!)
which would populate a few directories and leave it like that, empty.

We always overwrite the _manager attribute on the subsequent line so the
effective session we're using after  _commit_clean() returns is the last
one anyway.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>