Cherry-pick revno 3317 from lp:checkbox.
plainbox:exporter:xml: correct attachment and IO log handling

This patch re-writes how we handle control characters and text in both
the I/O log and attachment handling. First of all the root problem is
that we were operating on *bytes* and the regular expression that was
filtering out control characters was actually filtering out parts of
UTF-8 encoded non-control characters with code points > 127.

The first part of the solution removes CONTROL_CODE_RE_BYTES and patches
everything that used to use it to handle text. This involves the
_flatten_io_log() helper method. To quote the comment there:

This implementation attempts to decode the stream as UTF-8
(striping out everything that isn't UTF-8), filters out Unicode
control characters so that they don't appear in the session data
subset anywhere and saves the stream as a long BASE64 encoded
string, representing the same raw, underlying UTF-8 text.

The implementation is a little bit tricky, attempting to consume as
little memory as possible by decoding the string incrementally using
codecs.iterdecode().

The second part of the code that used to use CONTROL_CODE_RE_BYTES is
the now-removed overridden implementation of the _build_attachment_map()
method. The whole idea of overriding that method was so that we could
strip out the control character bytes but this was wrong for two
reasons:

1) we cannot strip anything at byte level, see above for the rationale
2) we should not clobber valid binary data (e.g. JPEG attachments)

So the whole overridden method is removed. Instead, wherever we need to
handle attachments *as text* we now use the new _as_text() helper method
which does a base64 decode of the raw bytes followed by text decoding
(via UTF-8 + 'ignore') and control character removal using the familiar
regular expression. Attachments that need to be binary are handled wit
the new _as_b64() helper method.

All of the places that touched data['attachment_map'] now use one of the
new helpers. In addition the _as_text() method was similar enough to the
as_text() local function to combine the two so dmi, sysfs and udev now
get handled the same way.

UPDATE: after reading the code again I realized that all the places that
assume stdout/stderr of any job can be decoded safely as UTF-8 need to
be patched so I've patched _add_questions() to use .decode("UTF-8",
"replace") to make sure it doesn't raise UnicodeDecodeError on
incompatible input.

Fixes: https://bugs.launchpad.net/plainbox/+bug/1375724

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>