plainbox:commands:run: split into cmd+inv
This patch splits the 'run.py' module into two new modules, cmd_run and
inv_run. During plainbox startup only the cmd_run module needs to be
imported and it has minimal requirements so very little overall code
needs to be loaded.

When the run command is actually executed then the inv_run module is
imported, bringing in all of the additional modules it requires.

Tests are adjusted to use the appropriate new module instead, as is the
'checkbox cli' command.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>