plainbox:commands:checkbox: split into cmd+inv
This patch splits the 'checkbox.py' module into two new modules,
cmd_checkbox and inv_checkbox. During plainbox startup only the
cmd_checkbox module needs to be imported and it has minimal requirements
so very little overall code needs to be loaded.

When any commands that actually needs to use the base 'checkbox'
functionality is executed then the inv_checkbox module is imported,
bringing in all of the additional modules it requires.

Tests are adjusted to use the appropriate new module instead, as are the
'checkbox sru', 'checkbox cli', 'plainbox run' and the 'plainbox dev
special', commands.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>