plainbox:commands:check_config: split into cmd+inv
This patch splits the 'check_config.py' module into two new modules,
cmd_check_config and inv_check_config. During plainbox startup only the
cmd_check_config module needs to be imported and it has minimal
requirements so very little overall code needs to be loaded.

When the 'plainbox check-config' command is executed then the
inv_check_config module is imported, bringing in all of the additional
modules it requires.

In addition, all call sites that used to import the old module are
updated to use one of the new modules instead.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>