plainbox: be lazier about loading providers
This patch changes the plainbox command in a few fundamental ways. First
of all the 'plainbox --providers' option is removed. It was deprecated a
good while ago and nobody used it anymore. Instead plainbox can be
invoked as 'stubbox' to load the stub provider. In normal mode (when
invoked as plainbox) stubbox is *not* loaded anymore. Some commands
still refer to themselves as being a part of plainbox, 'plainbox dev'
for example, but this is hard to change and not important enough.

This is the only user-visible change here.

The other part of this patch moves the initialization of
PlainBoxToolBase._provider_list far later into the startup process, to
the dispatch_command() method. This is not the ultimate goal as we still
load providers too early (before we even know we need them) but it is
the least invasive (API wise) change that we can make now.

Overall the startup cost of 'plainbox --help' is now much lower.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>