plainbox:runner: fail gracefully on windows
This patch changes the JobRunner to fail gracefully by returning the
OUTCOME_NOT_SUPPORTED result when none of the execution controllers is
capable of running the requested job.

This change affects a private method that is unfortunately called by two
other places, the oldcli module in checkbox-ng (deprecated, will be
removed soon) and the script command that should be soon rewritten on
top of SessionManager alone. None of those places handle LookupError
that can be raised when the current device cannot execute a given job.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>