{plainbox,checkbox}: load providers lazily
This patch eradicates the concept that provider_list is loaded somewhat
early in the startup of all command line tools and then gets passed around
everywhere it might be needed.

This is wasteful (as not all commands even need to know all providers)
and makes the tool seem less responsive by delaying the time first output
is printed by a noticeable amount.

Instead, a callable gets passed and calling that callable for the first
time actually triggers the loading. Subsequent calls are free but nobody
is calling this more than once anyway.

Several pieces of the code were adapted, specifically all commands just
pass this through while all invocations (the other part of a command
that is used when the command is actually asked to perform something)
were taught to store the loader and provide the same 'self.provider_list'
lazily.

Several tests were updated to pass a callable instead of a list but
otherwise nothing major is being changed.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>