checkbox-support: add CRASHED / 'crashed' test status
This patch adds a new test status (aka outcome) mapping so that
'crashed' is forwarded as is from the XML report to the parser's caller.

The intent is to use the new status as a way to convey the information
about a malfunctioning test, as opposed to a test that is merely
failing.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>