plainbox:command: make Colorizer smarter by interpreting color as policy
If the Colorizer is initialized with True, it will always use colors.
If the Colorizer is initialized with False, it will never use colors.
If the Colorizer is initialized with None, it will use colors if sys.stdout
is a TTY. Otherwise, it will use whatever was passed as argument (keeping
compatibility with current APIs)

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>