plainbox:exporter:text: fix TextSessionStateExporter to handle option_list
Thechnically each exporter has an option list argument that it must handle
int its initializer. This wasn't really tested that often but it got broken
when we added simple color support to the text exporter.

This patch restores the ability to pass an option list, moving the,
optional, color argument to be the second argument.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>