plainbox:session: look for non-hidden .session directories
This patch changes how plainbox discovers and enumerates sessions.
Previously any directory would be a "session" that would be inspected.
With this change only non-hidden directories that end with '.session'
are considered. This makes it friendlier to work with when storing
sessions in a different location, such as the regressions repository.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>