plainbox:vendor:extcmd: defer on_{,abnormal_}end() until I/O stops
This patch fixes race condition inside extcmd, where a process can exit
but we can still see I/O on the pipe _after_ we saw the SIGCHLD signal.

This can happen if the pipe is leaked to a third process or if the
particular ordering of events (theorethical) happens to be that we
fierst read the signalfd() descriptor and only after we read the
(perhaps last words of) dta from the pipe.

Fixes: https://bugs.launchpad.net/plainbox/+bug/1402710

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>