plainbox:vendor:extcmd: change algorithm for reading trailing pipe data
This patch fixes a potential bug where plainbox won't see the trailing
data from a pipe if a given scheduling constraint happens to put the
death of the process _before_ the notification that we have some more
data pending.

The new algorithm lets each pipe to finish naturally but enables the
special mode force_last, where we read from a pipe at most once before
closing it unconditionally, if we already saw the deadth of the process.
It has the same guarantees as the old algorithm but it will (unlike
before) also work if we see the death before being notified about the
last data.

I don't know if this is realistic but I'm much happier with this
prevented non the less. It doesn't seem to happen anymore now that we've
fixed pipe size bugs that made it happen all the time.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>