checkbox-ng:service: emulate JobDefinition.via for D-Bus
This patch adds a way to emulate the deprecated ``via`` attribute so
that it can be retained "as-is" in the current D-Bus interface. The
approach is rather crude, since we only allow for one session anyway, a
global session object is used to look up the job state associated with
the job whenever JobDefinition.via is accessed.

Tested-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>