checkbox-ng:service: fix .via for jobs not in session
This patch fixes a crash when JobDefinitionWrapper.via was being
accessed after resuming a session, when the corresponding job was not
being a part of the session (but is still available on DBus via a
provider).  For example:
 - create a provider with jobs [a, b]
 - create a session with job [a]
 - save the session before a runs
 - resume the session, the session now has [a] but not [b] in it
 - attempt to access b.via

In the current code, this would crash as there was an implicit,
incorrect assumption that each job visible on DBus is also a part of the
current session. The patch corrects that assumption by allowing no
JobState to exist for any JobDefinition.

(A copy of the relevant comment on the bug corresponding report)

1) Job definitions are either generated (and thus bound to a session) or
   not (and thus bound to a provider)
2) The .via property was previously a mutable state on a job. Now it is
   something one can access from session-bound JobState.
   2a) this means that .via cannot be accessed before a session exists
   2b) this means that .via cannot be accessed if a session doesn't
       contain a given job (despite that job existing).

In the past, via was always accessible, perhaps holding nothing or
perhaps holding stale data (from previous session run).  Currently when
2b happens, we crash.

The fix is to simply return "" when there is no session or there
is no corresponding JobState in the session.

Fixes: https://bugs.launchpad.net/checkbox-ng/+bug/1412660
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>