plainbox:secure:plugins: add LazyFileContent
This patch adds a simple class that can pretend to look enough like
a string representing the content of a loaded text file to fool all of
plainbox. The actual content is loaded the first time it is needed.

The trick is used to prevent loading files we don't intend to analyze,
such as arbitrary, large binary data files, while still using the illusion
of a single plugin-based collection interface that is friendly for testing
/ mocking and further evolution.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>