plainbox,checkbox-{ng,touch}: use new provider properties
This patch replaces the use of method-based provider APIs such as
get_units(), get_builtin_whitelists(), etc, with the eaier, property
based APIs.

The following substitutions are made:
  p.get_builtin_whitelits => p.whitelist_list
  p.get_units()[0] => p.unit_list
  p.get_units() => p.unit_list, p.problem_list
  p.load_all_jobs()[0] => p.job_list
  p.get_all_executables => p.executable_list

In two cases, a for-loop over all the units in a provider was replaced
by p.id_map-based lookup. This reduces O(N) complexity to O(1). While
the right-hand-side of that expression is a list, that list will always
contain one element if a provider is valid.

At the same time, some test code is updated to mock / fake the properties
instead of the methods. Depending on the code, id_map may need to be mocked
as well.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>