plainbox:xparsers: exclude lineno and col_offset from repr()
This patch adds a custom __repr__ to the base Node class in order to
shorten various doctests significantly. This removes the means to test
lineno and col_offset this way but it's a small price to pay for clarity
and we're not currently using col_offset anyway. By the time we do we
could test both in a different way.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>