plainbox:session: add note about complexity
This patch adds a small note on how to improve the
_bulk_override_update() method. The algorithm I was thinking of is based
on combining *all* regular expressions that cause any overrides into one
large regular expression (using the | operator) and associating a set of
overrides with each accepting state.

Using a properly crafted algorithm, the regular expression could be
matched to each word in parallel (so that all edges are followed) if
this is permitted (this is essential for matching wildcards) and
combining the union of sets stored in all reached accepting states.

This would transform the loop from:

    for pattern in patterns:
       for job in jobs:
           if job matches pattern:
               apply overrides associated with pattern

To something like:

    combined_pattern = combine(patterns)
    for job in jobs:
        if job matches combined_pattern:
            apply overrides associated with patterns that matched

This is trivial to do (as in, using built-in regular expression engine
from python) in the naive case where at each job can be matched by at
most one pattern. In the realistic case where this limitation doesn't
apply a different way of matching patterns must be used (as explained
above).

I think that it still provides improvement in complexity as the
combined_pattern state machine can be minimized to combine states and
that the complexity is proportional to the number of edges. In the
original implementetion this would have been still proportional to the
number of regular expressions and I think this value can be larger.

One (separate) improvement would be to combine all of the known job
identifiers into a suffix tree and run the algorithm on that tree. Then
we can reduce that to O(1) given that we have the prefix tree (could be
cached by a provider if we also support handling generate jobs
separately) and ignore the linear complexity in the number of edges in
the resulting combined_pattern.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>