plainbox:unit:job: support jobs with arbitrary IDs in depends:
This patch changes how we parse the depends: field. In the past it was
"parsed" with a simple re.split('[ ,]+') which was okay-ish. Now it's a
bit problematic when generated jobs can indlude arbitrary characters and
for other reasons (hexr limitation) we want to include this data in the
actual job id, not only the summary field.

The solution is to use a simple parser (WordList) that is based on our
only scanner (WordScanner). Unlike in the past, it does have one error
case, using the equals sign will trigger a non-fatal parser error that
logs a (imprecise) warning about junk.

To avoid breaking the API now I'd like to address that in validation
later.

Fixes: https://bugs.launchpad.net/plainbox/+bug/1425977

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>