plainbox:provider_manager: use new validation by default
This patch switches how 'manage.py validate' works by default. For a
good time we had the -N option, aka the --new-vaidation-core, which used
a totally separate, much better validation subsystem. It's now time to
make that the default and get rid of the old validation code entirely.

This patch just flips the default and hides the option (keeping it for
backwards compatibility before it gets removed later in the future).

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>