plainbox:providers: add the manifest provider
This patch adds a new special provider for working with manifests.  The
provider is in the plainbox name-space and might be handled differently
by the core or by application later. The provider ships two jobs, one
that converts a manifest into a resource (a resource job) and one that
creates the manifest (a user-interact job).

Both of those jobs might be integrated into the flow of a testing
application later. For now, since we tightly control their identifiers,
it is safe to use them as any other jobs (just add them to a test plan
and you are done).

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>