plainbox:ctrl: rewrite tests for local jobs
This patch rewrites three tests for local jobs, not changing their
semantics but changing the implementation. The original implementation
used mocking for everything and focused on observing impact of the
session controller on the mocked session. This turned out to be difficult
to adapt to changes to the controller so instead non-mocked objects are
used and minimal mocking is used as a short-cut (to remove the necessity
to create real results that correctly define jobs).

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>