plainbox:suspend: add 5th suspend format
This patch adds the fifth suspend format. This format is the same as the
fourth format with the exception of how DiskJobResult is saved. If the
new (but still optional) session_dir argument is provided then a
relative pathname to each DiskJobResult is saved instead of an absolute
one. This allows the whole session to be relocated easily.

Fixes: https://bugs.launchpad.net/plainbox/+bug/1378295

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>