plainbox:testplan: don't crash if id is not set
This patch prevents a crash that can happen when a test plan doesn't
have an ID field. This is normally validated separately but the context
validator (which does other checks) should be written defensively and
not assume that each unit is correct.

Here the problem was related to trying to come up with a regular
expression match of a None id against some unrelated test plan pattern.

Fixes: https://bugs.launchpad.net/plainbox/+bug/1443899

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>