checkbox-support:parsers:submission: Add parsing of kernel_cmdline_attachment.
This includes the kernel commandline parser (trivial since we don't actually
parse the commandline, we just extract it verbatim from the raw attachment data)
as well as the changes to SubmissionParser and the TestRun class to actually invoke
the subparser when needed and gather the data.

Testing changes add the xml chunk data file for the test, the test
method itself, and augment the SubmissionRun data-holding class with a
setKernelCmdline method which we'd expect the submission parser to call
when it finds and parses a kernel_cmdline_attachment.