checkbox-support:parsers: Use filenames instead of file-likes for submission parser testing.
This caused problems with a test fixture which included strange non-ascii (but xml-valid) characters.
Since etree.parse was being passed a file-like object, it was unhappy with the encoding/decoding and
was giving UnicodeEncodeErrors.

etree.parse supports passing a filename, in which case it handles opening and encoding/decoding the file and is happy.

This affected only testing under Python 2.7. The tests were verified under both 2.7 and 3.4 and they worked OK.