plainbox:result: add JobResultBuilder
This patch adds a new class that assists in building appropriate
JobResult objects. The result builder has a .get_result() method that
returns a regular result (of the appropriate kind, either disk or
memory). The idea is to make the result immutable later on and only
allow modification to happen inside the builder.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>