plainbox:session:assistant: rewrite the SessionAssistant
This patch largely rewrites the entire session assistant codebase. The
intent was to always make the SessionAssistant capable of replacing all
the code we use in all the applications based on plainbox. This is the
first version that can attempt that.

Many design choices were revisited, instead of somewhat weird generators
returning "events" we now have morris-based signals everywhere. The API
is far more featureful, catering to the needs of more applications.

All of the methods have better documentation and tests to ensure that
they really work as expected. In addition, with the advantage of the
UsageExpectations system application developers cannot stray into
unknown paths by executing methods in unexpected order.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>