Initial import from lp:checkbox r3943
This patch moves the Canonical certification client provider away from
the lp:checkbox common tree. The reason for this is rather trivial, the
releasectl script doesn't support it correctly at this time and it would
be much easier to support it separately.

At this time the huge changes of the "early days" are over and it should
be quite okay to let it live separately.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>