providers:checkbox: Adds category ids to all jobs (except local jobs)
This patch adds category ids to all jobs in the checkbox provider (almost all
jobs) so that users can select the category they want to test in
checkbox-converged.

local jobs are not getting a category id as those will be run during
bootstrapping. Legacy jobs in install, panel-reboot, panel-control and
server-services are also kept uncategorised.