plainbox: add support for 'after' constraint
This patch adds the 'after' ordering constraint. It is exactly the same
as 'depends' but unlike it, it will not inhibit a job from running if
the other job runs but fails (or crashes). It is especially useful for
enforcing the order on a pair of jobs where the job that runs second can
collect the log files or the system state after the execution of the
first job.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>