plainbox:session:assistant: use tuples for creating controllers
This patch changes the signature of the function responsible for using
alternative execution controllers. Now the funciton takes an iterable with not
only the class of the controller to use, but also params that are forwarded to
the class' __init__ mehtod. This way Controllers still can be instantiated in
the session assistant territory without applications needing to import
plainbox' implementation details.

Signed-off-by: Maciej Kisielewski <maciej.kisielewski@canonical.com>