plainbox: expand SessionAssistant initializer API
This patch modifies the session assistant initializer to allow up to
four arguments. The three new arguments are:

app_version: so that we can use this implicitly in some places,
 e.g. don't resumes sessions created by future versions, etc.
api_version: so that we can change usage expectations over time
 but let applications stay compatible by using a fixed API version.
 This can be changed to a __new__ call that returns a versioned
 SA class instead of doing if-then-else magic in all the places.
api_flags: so that we can allow applications to opt-into optional
 features and so that we can adjust expectations accordingly. This
 will also allow us to easily compare applications for feature
 parity.

For now all new arguments have sane defaults. Once all applications are
patched the defaults will go away.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>