Add an option to use fwts to run power management tests
Power management testing script (pm_test) was based on another script
(sleep_test) to perform suspend and resume. This was the cause of problems on
some devices.

Added a new optional parameter (--fwts) to pm_test that enables to
use fwts_test (a wrapper to fwts) when testing suspend. This uses values provided
by the QA team.

Removed time check in pm_test as this is not needed for certification (there is
already a job for that).

Finally, modified stress job `suspend_30_cycles_with_reboots` to use the new
parameter.