plainbox:ctrl: catch tempcwd errors and print warning
This patch handles PermissionError exceptions coming from TemporaryDirectory.

Consider a job, that is run as root that runs cp somedir .

When the temp cwd context manager is left, the cleanup is run as the user
running plainbox, not root. This caused plainbox to crashed.

This patch makes plainbox display a warning and proceed.

Fixes: https://bugs.launchpad.net/plainbox/+bug/1569575

Signed-off-by: Maciej Kisielewski <maciej.kisielewski@canonical.com>