providers: properly generate whitelist patterns from test plan's include
Whitelist qualifiers are generated with a code like:
 for pattern in pattern_list
The pattern_list that was supplied was (in case of 'guessing whitlist from a
test plan') test_plan.include... which is a string!
That meant that str was treated like a list, and for longer include fields,
many thousands of those regex qualifiers were constructed. Aweful!

This patch repairs how those patterns are generated by parsing include field,
enveloping them with a historical ^ and $, and pushing them to the whitelist.
The result is 25% faster plainbox spool up.
Signed-off-by: Maciej Kisielewski <maciej.kisielewski@canonical.com>