launcher: fix crash when ui==silent and no secure_id is provided
In a fairly skimpy launcher with [ui]\ntype=silent I stumbled upon a problem
where checkbox would crash because of a missing secure_id. Missing, because UI
wasn't interactive and there was no way to ask user for it.

This patch detects such combination of events, and skips the creation of that
default stock_report.

Signed-off-by: Maciej Kisielewski <maciej.kisielewski@canonical.com>