parsers:udevadm: Run lsblk from parse_udevadm_output
If the parse_udevadm_output lsblk argument is None then run the `lsblk` command
in a subprocess.

This is required since we use plainbox dev parser udevadm to attach the outout
as a json blob in the sumission report.

To be still compliant with the way unit tests are written, tests will pass
an empty string instead of None.