make audio_settings.py rely only on pactl
Switch to pactl from pacmd required change in invocation of 'list devices'
and because pactl list formats output a bit different, so one regex also got
amended.

audio_settings controls the pulse server, not reconfigures it, so following
manpages, pactl is more suited to do it. Also, in pulseaudio-snappy, there is
only pactl.

Signed-off-by: Maciej Kisielewski <maciej.kisielewski@canonical.com>