Send all data to stdout/stderr at once to avoid later decoding errors
When displaying a testing session results on screen (stdout/stderr), the
StreamTransport class copy the raw data to a ByteStringStreamTranslator that
then decode it into Unicode text in order to show it properly on screen.

The method used was using a buffer and sending 16 KB chunks, which sometimes
lead to a UnicodeDecodeError exception because a Unicode character would be
spread between two chunks of data, making its decoding impossible.

For instance, the Unicode character ☑ is composed of 3 bytes (0xE2 0x9C 0x93).
If this data is cut and only the first part (0xE2) sent in the data chunk, it
is impossible to decode it into a known Unicode character and plainbox crashes.

To avoid this, we make sure the whole data is sent at once.

LP: #1663166