certification: Create a dedicated SubmissionServiceTransport
Reuse of CertificationTransport is not possible as the new submission service
does not expect a form but a binary blob.

The difference is that using data is not encoding the body like a form,
but using files makes it a form.

Note: When the submission succeeded the path to the url is given in 'status_url'