Make some tweaks on some existing test cases and add three more test cases.
1.As of now, people either need to add sudo or switch to root to run docker command
on Ubuntu Core 16, The reason here is that system user and group can not be editable
as /etc/[passwd, group] is read-only on Ubuntu Core. We need to support to run docker
command without sudo once the new interface(user& group management) is supported on snapd.
2.Pipe a snapcraft.yaml from STDIN to compose doesn't work in docker.compose 1.12.
0 which is the one we're using to snap docker docker_17.03.1-ce-1.
It's officially fixed in docker compose 1.13.0.
We'll use latest docker compose for the next docker snap release once the upcoming quarterly
release of docker(17.06-ce) is out. As a workaround, we generate docker-compose.yaml file locally
and let docker read from the file to compose an image.
3.Add three more unit test cases for coverage purpose.