handle instatiation in `checkbox-cli run`
Before the session is finalized, SA's hand_pick_jobs is called with the same
pattern as in the first run. If any new job is matched, that hasn't run yet,
the whole main-job-running-loop is run. This happens until there is no new
job generated. This means there can be (theoretically) an infinite number
of steps - e.g. new instance instantiating new instance, and so on.

I think this makes plainbox' PXUs turing complete...

Signed-off-by: Maciej Kisielewski <maciej.kisielewski@canonical.com>