override the driver name to amdgpu-pro only for vendor == amd
previously, if the system had vulkan-amdgpu-pro pkg installed, the script
reported 'amdgpu-pro' as the driver for all the GPUs.

Now it's amdgpu-pro is reported only for GPUs from vendor 4098 - AMD.
This is still not perfect, as theretically there might be multiple amd gpus
in the system, not all of them amdgpu-pro, but it still makes it closer to the
truth and right for the typical intel+radeon configuration.

This patch also rearranges the logic for the dpkg-query call.
if not `some commands run` intuitively should mean "if the command failed"
which in the case of subprocess.call(...) is casted to boolean, therefore
`if not subprocess.call(...)` means success. I added explicit comparison to 0.

Fixes: LP:1731660

Signed-off-by: Maciej Kisielewski <maciej.kisielewski@canonical.com>