launcher: Add the --message argument to provide a submission description
The content of the message can be a multiline variable. It will be stored
in the session metadata app_blob and recorded in the json report as a new
description field for C3.

For consistency, checkbox-cli submit gets the same feature. The trick to
support it though requires extraction of the submission archive to modify/add
the description data and a generation of a new archive in a temp dir.