allow jobs to leave outcome info for auto-resume resolution
Previously when checkbox was auto-resumed (e.g. when running reboot stress
jobs), the outcome was always decided to be passed.

This patch introduces a way for jobs to communicate back what the outcome
should be. This way if a job detects a problem with a job, it can leave
appropriate message in the CHECKBOX_DATA dir.

The feedback file that checkbox looks for is __result. It should be a json
object containing fields that will be used by MemoryJobResult, most important
of which is the "outcome". Other useful field is "comments" so the job can
give verbose information about what went wrong.

If the file cannot be parsed, or the file isn't there checkbox defaults to the
previous behavior of marking as passed.

Signed-off-by: Maciej Kisielewski <maciej.kisielewski@canonical.com>