session: optimize creation of throwaway session managers
Instead of creating one every time it's needed, "cache" them in the
SessionManager cls. The cache is keyed by the providers requested.

Those managers are needed when validating test plans. For each nesting TP
new session is created (and all providers are parsed; all jobs created).
They are also used when the outcome of the boostrapping job was handled.
On slower machines it meant that validating plainbox-provider-checkbox
took 1m28s. With the caching mechanism introduced by this patch, the time
dropped to 0:36s.
As for bootstrapping gains, going on checkbox-plano from CLI to test-selection
using .caracalla-test-runner, it went from 1m41s to 1m15s.