bin/run-all-tests-locally.sh: add dev script for running tests locally
This facilitates local developer testing of a built snap using these
scripts, and also will cache the stable and delhi channel revisions for
future tests that ensure refreshes from those revisions to the revision
under testing work.

Also enable all the current tests to work with this script by checking
for an env var and if it's there then install that snap revision instead
of always using beta.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>