execution: use sane defaults for runner configuration
Default __init__ of the UnifiedRunner left some of the fields set to None,
defering "proper" initialization to the caller/user. This made checkbox crash
on most uses that didn't do the later initialization.

This patch moves some password and user-change related code closer to the core
so they can be used directly by the UnifiedRunner __init__.

Signed-off-by: Maciej Kisielewski <maciej.kisielewski@canonical.com>