Add TPM PCR7 unsealing automated test
This job checks that, once the PCR7 is extended, we cannot unseal data
using it. A reboot is necessary to reset the PCR7 to its original value.

Note that this may impact the TPM2 automated test suite. Investigation
is required to check if this job needs to be run *after* the TPM2 test
suite or if we should make sure there is a reboot between this job and
the TPM2 test suite.