snap_utils:system: Add get_lk_bootimg_path()
Helper function used in many downstream scripts to get the path
of the booted img partition on system using the little kernel as a bootloader.

Use the kernel_cmdline parser output to pick the parameter snap_kernel=<name>.
The <name> value should be then be used to select the correct booting_matrix
line in snap_boot_selection.