test-sys-mgmt-agent.sh: update for fuji
The last test case sends a start service message
for an invalid service, and checks curl for a non-
zero retval. This no longer works in the Fuji (1.1)
release of EdgeX as the SMA now returns a JSON reply.
This commit updates the test to check Success field
of the returned response instead.

Signed-off-by: Tony Espy <espy@canonical.com>