assistant: better info when transport fail due to exporter's error
If the call to send results fails b/c of an exporter failure (like when tar
exporter fails due to a bad json character) the exporter exception got
propagated, but there was no info that the submission to certification
site failed.

This patch makes sure the warning about transporter failure is printed.