session:assistant make manifest path user agnostic
The assumption made that manifest path should in the home dir
of the current user is incorrect. With support for "normal_user"
the manifest collection job will be run as this user and the
home directory will no longer match. The simplest approach to
solving this is to make the path user agnostic.