Remove sys-mgmt-agent from active/enable services
The sys-mgmt-agent is disabled by default in Geneva due to
the fact that it has a hard-code service list. Since the snap
by default disables all the support services, the SMA continually
logs errors every 15s due to these services not being available.

See the following issue for more details:

https://github.com/edgexfoundry/edgex-go/issues/2486