parsers:udev: Consider all /dev/dri/card* as valid VIDEO devices
On arm devices, identifying VIDEO controllers can't be done from usb/pci
modaliases or classes. And most of the time none is reported.

Though, when the kernel creates a /dev/dri/card, it means a video controller
is enabled and can then be categorized as a VIDEO device.

To avoid creating both a VIDEO and a DRI device, the following rules are applied:

1. When there's a least one VIDEO category found by the udev parser, just ignore
all DRI devices and remove them from the list of known devices.

2. If no VIDEO devices are found but some DRI are, then re-categorize them
as VIDEO. It will allow existing graphic jobs in the checkbox provider to
work properly and jobs using templates to be generated.

Note, as a product name the ID_FOR_SEAT property is used.