resource: Don't raise AttributeError when parsing resource expressions
Problem is that doing so the full expression is evaluated as False.
Like this one:

package.foo == 'bar' or valid_resource.valid_attr == 'yes'

Even if the second half could be true, if foo doesn't exist an exception
is raised and such requirement will never work.

In such case returning the empty string will cause the first half to be
evaluated as False which is by far the most logical evaluation.