pm2.0_4.1.1: modify some tests for the platforms without sha1 pcr band
Most platforms which default don't support sha1 pcr band and can be
changed on the bios setup menu. Sha1 pcr band is not mandatory from
the TCG TPM spec. So, modify some tests for not getting lots of failures
on those platforms which sha1 pcr band is not default supported.

* helpers.sh: add helper function for check sha1 pcr support
* skip/modify tests related to sha1 pcr band.

Signed-off-by: Ivan Hu <ivan.hu@canonical.com>