Fix: remove version spec of xlsx from setup.py
This patch should help us with future PMR hiccups.
according to pwlars:
"""
I think I might have found something - take a look at
this I wasn't able to get it working by installing
python3-setuptools-scm deb package, but since we definitely know that
there are a LOT of versions of xlsxwriter>=0.3, then the >=0.3 is a bit
redundant. If you take the version spec out, it seems to work fine.
"""