Fix: don't crash on no buildstamp info
This fixes a tarball exporter crash happenning when buildstamp job yielded no
output. Now, in this circumstance the exporter will have "Unknown" in the
buildstamp field. With that being said, this *should* no longer happen as
there is a corresponding patch to the buildstamp job that makes the job fail
on no info.

See LP: #1933632 for details.