Change: bin/kernel_taint_test.py: move report_failures() to main() and code cleanup
The report_failures() is now just return whether the test pass or fail
move the code piece to main() and update the docstring accordingly.

The tainted bit value and its message will be printed at the beginning
of this if statement, there is no need to print it again.

Remove unnecessary parentheses for return

Use raise SystemExit() for taint_file not found exeception.