Remove: cruft code @ c3 transport
For quite some time the only way the send method of a submission transport was
used was without any params. The secureID is established in the __init__ and
the config is never read by both: normal session running or the `submit`
subcommand.  The configs that were declared there (and not used) use the old
PlainBoxConfig class that is not in use anymore.
This patch greately simplifies the code related to submitting to C3.