Change:fwts_test: print content of log file
The information contained in the log file is very useful for debugging,
yet it is currently buried in the submission archive.

In addition to saving the information in the log file, this commit
appends the content of the log file to the summary.

This approach was chosen because it will benefit all the Checkbox jobs
calling fwts_test.