Fix: traceback when launching checkbox
This patch adds an optional arg to detect_restart_strategy so the information
that the session is remote can be used to pick the best strategy.
There's a bit of heuristic in the SessionAssistant class that guesses that the
session is remote by looking at its title - remote sessions have
'checkbox-slave' in the session's title.

Fixes: LP: #1966045