Change: graphics/*_maximum_resolution_* compatible with X and Wayland
On Wayland, xrandr returns the current resolution as the maximum
resolution, even when it is not. For instance, if a laptop has a
3840x2160 screen but is currently set to 2560x1440, xrandr would
return 2560x1440 as the maximum available resolution.

Moreover, the python oneliner to check current resolution breaks on
recent versions of Gtk because Gdk.Screen has been deprecated.

Introduce a new script that:

- retrieves monitor(s) info from sysfs (port, max resolution, status,
etc.)
- uses up-to-date Gtk 3 methods to grab the current resolution and
scale factor for each connected monitors
- displays this information for tester to review
- compares data from sysfs and Gtk and returns an error if their data
does not match (so that the graphics/*_maximum_resolution_* job status
is automatically set to Passed or Failed depending of the outcome)

Update the graphics/*_maximum_resolution_* job to call this script.

LP: #1968943