1. Fix LP: #1983657
The dmi.product list does not include convertible chassis type in some of testing jobs.
It will skip the job automatically even if the system is supported with.
This problem might have a risk to missing the issue if the function does not work.
These items need to be modified and added the convertible type.