core-snap: get rid of checkbox-provider plugin
In order to enable building the series22 snap with snapcraft 7.x we
needed to ditch the checkbox-provider user-defined plugin from the
snap's source. This patch does that by overriding the build step of each
of the providers with a scriplet that does essentially the same thing.

One other change is fix of unbound variables in scriplets as snapcraft
7.x rejects build scripts with those.