support: update pactl parser to be compatible with pyparsing 2.x and 3.x
pyparsing 3.x, introduced in Ubuntu 22.10 repositories, removes support
for __compat__.collect_all_And_tokens setting [1]. As a result,
MatchFirst expressions return a list, even if it only has one item in
it.

Since Ubuntu 16.04 (which ships pyparsing 2.0) still has to be
supported, this commits makes use of lambda for some expression
fragments to work on 22.10 while being retrocompatible with 16.04.

[1]
https://pyparsing-docs.readthedocs.io/en/latest/whats_new_in_3_0_0.html#other-discontinued-features