Change: evdev-based touchscreen jobs handle timeout themselves
Previously, the evdev_touch_test.py had a --timeout parameter, but the
implentation within the script did not seem to yield good results in
recent versions of Ubuntu. It was removed and timeout is now handled in
the Checkbox jobs directly.

Therefore, if the Python script does not return anything within 10
seconds, the timeout command will return code 124 and the Checkbox job
will be marked as failed.

LP: #1974197