fix: make USB insertion not miss events in journal
The previously used `seek_tail` method moved the journal cursor beyond
the point where messages could go, so when the "mass storage inserted"
message came, it may have gone unnoticed.

This patch uses seek_realtime function that moves the cursor to
particular time ("realtime" here is a misnomer). In this case it
moves the cursor to the current time, so all journal messages from that
time on are beiing checked.