metabox: fix container creation (and other various fixes) (#272)
* metabox: add origin in string representation

MachineConfig.__str__() is used as the container's name when creating
it. Some scenarios could use the same role and alias, but different
origin. In that case, metabox would fail when trying to create a
container that already exists.

Adding the origin to the MachineConfig string representation prevents
this.

* metabox: change origin naming from `classic_snap` to `classic-snap`

LXC container name can only include alphanumeric characters and
hyphens.[1]

Since origin is used in the MachineConfig string representation, and
since this is in turn used when creating the container, the underscore
is replaced with an hyphen.

[1]
https://linuxcontainers.org/lxd/docs/master/reference/instance_properties/#instance-properties

* metabox: change default snap name to `checkbox`

checkbox-snappy and checkbox-snappy-classic snaps have been deprecated
in favor of the checkbox snap.

* metabox: follow branches and snap naming convention for the checkbox snap

The snap track map for the classic snap now follows the release number
(e.g '20.04' instead of '20').

There are also a new track and a new snap for Jammy.

* metabox: add debug logging info during container init

* metabox: typo