Ask tester about available graphics ports and adapters instead of relying on resource job (#282)
* Add manifest entries for graphics ports

* Split USB-C manifest entry

Some devices under test (DUT) have one or more USB Type-C connectors.
Some of these connectors offer direct HDMI/DisplayPort outputs if user
has an appropriate cables, while for others, a USB Type-C to
HDMI/DisplayPort adapter can be used to connect the DUT to an external
monitor.

For these cases, a new manifest (`has_usbc_video`) is created.

The old one, `has_usb_type_c`, is renamed `has_usbc_data` as it's used
for all tests related to data, be it USB drives, Ethernet adapters or
HID (keyboard, etc.).

* Update jobs requiring `has_usbc_data` manifest

* Use graphics ports manifest entries to decide whether jobs should be run

Instead of relying on the display resource job (that relies on
xrandr/gnome-randr to determine heuristically support for different
connectors), use manifest entries to determine if the monitor and audio
jobs should be run at all. Since manifest entries are decided manually
at the beginning of a test session, it will be more accurate.

Some very old jobs (monitor/{index}_svideo_{product_slug} and
monitor/{index}_rca_{product_slug}) that are not used in our current
test plans are kept, but the requirement on the display resource job is
removed so the display resource job can be removed entirely.

In addition, now that there is a dedicated manifest to handle video
output through the USB Type-C connector, a lot of jobs are updated to use
it instead of the older `has_usb_type_c`.

Fix #56

* Remove display resource job

This job was not reliable since it was based on heuristics that often
failed, especially on 22.04.

Replaced by manual manifest entries that are more accurate.