Modify 30 cycles reboot/poweroff attachment jobs names
Since attachment jobs generate a file named after their id, it is not
clear what `suspend-30-cycle-log-attach-with-reboots` contains when
opening the Checkbox submission.

For testers and developers, the names `pm_test.reboot.3.log` and
`pm_test.poweroff.3.log` make more sense, so in order to bring more
clarity when debugging, these are used as suffixes for attachment ids.