Add manifest for touchpad-related jobs (#293)
The original jobs appear and be executed based on the Chassis type from
DMI information. However, there are more and more IoT devices which
don't equip the physical touchpad nowadays. And for some reasons, their
Chassis type matches the condition of the touchpad job, this causes those
touchpad jobs will be executed but tester need to skip them manually.

Therefore, the manifest of touchpad will bring a huge benefit for us to
filter jobs out in advanced.

I add the has_touchpad manifest option and apply it to all of the
touchpad jobs. At the meanwhile, I remove the dmi.product requirement.