Rework the Source Container Machine
The ContainerVenvMachine class used python virtual environments to
install Checkbox inside a container. It is renamed
ContainerSourceMachine and now deploys Checkbox directly into the
container using pip.

It now handles the creation of a systemd service, so that the service
container can be restarted, it will become reachable as soon as the
container is up.

It also copies over the following providers, so they can be used in
scenarios:

- base
- resource
- certification-client

Note that since Checkbox is now a monorepo, the `uri` parameter of the
configuration file has to point to the root of the Checkbox monorepo and
not to the checkbox-ng directory.

The configuration to use the source container is:

configuration = {
    'local': {
        'origin': 'source',
        'uri': '/path/to/checkbox',
        'releases': ['focal', 'jammy'],
    },
    'remote': {
        'origin': 'source',
        'uri': '/path/to/checkbox',
        'releases': ['focal'],
    },
    'service': {
        'origin': 'source',
        'uri': '/path/to/checkbox',
        'releases': ['focal'],
    },
}