Upload snapcraft logs as an artefact if checkbox-core snapcraft build fails
Snapcraft 7.x stopped outputing everything to stdout and instead puts
build logs in a `~/.cache/snapcraft/log/` directory. This is not ideal
when running snapcraft as part of an automated workflow, because errors
are not visible in stdout anymore, instead simply stating:

Logging execution to '/home/runner/.cache/snapcraft/log/xxx.log'

Add a step to capture these log files as artefact and upload them to the
Github Action if the build fails.

Fixes CHECKBOX-445