Require user to provide comment for manually skipped or failed cert-blockers
When issuing certificates for a device, the team reviews the different
test reports generated. Jobs that have a certification status set to
"blocker" must pass in order to issue a review.

Sometimes, it is not the case, but users do not necessarily provide a
comment that would help with the review.

Semi-automated jobs (`user-interact` and `user-interact-verify` [1]) can
be skipped before being started or after they've run.

Manual jobs can be skipped after been "run", meaning after their content
has been displayed to the user and the user has to pick an outcome.

These are handled in

- checkbox_ng.launcher.stages._run_single_job_with_ui_loop
- checkbox_ng.launcher.stages._interaction_callback

Checkbox local and Checkbox remote call _interaction_callback()
differently; Checkbox remote does not provide the current job's state,
which is required to know the job's effective certification status (this
status can be set in the job definition, but overriden in the test
plan). checkbox_ng.launcher.stages._interaction_callback() signature is
modified to pass the job state.

Fix #387

[1] https://checkbox.readthedocs.io/en/latest/units/job.html#job-fields