Modify session resume behavior on Checkbox local
When a session is resumed, Checkbox (local mode) will ask the user what
to do.

If the previously run job is a cert-blocker, a behavior similar to a
normal job run is implemented: if the user wants to set the job outcome
to `skip` or `fail`, they will have to justify why by adding a comment.

In order to do so, a "Add a comment" choice is added, and a loop is run
until all the conditions are satisfied.

For non-cert-blockers, an automated comment is added depending on the
choice, except if a comment was manually entered.