Make sure all the expected snaps are built during the release process
`snapcraft --remote-build` may fail to build all the snaps required (one
per architecture), and yet return 0 (no error), resulting in a "passed"
GitHub workflow even when not all the snaps have been submitted.

To prevent this, add a step to check the expected number of snaps is
present.

fail-fast strategy is set to false so that even if one of the workflow
fails, the others will continue. This is useful because it is possible
to manually re-run the failed workflow (for instance, the one for
checkbox22), so the others (checkbox16, checkbox18, checkbox20) should
not be impacted.

[1] https://github.com/snapcore/snapcraft/issues/4142

Fix CHECKBOX-446