Stop building HTML, man pages and translations when building checkbox-ng
With the move of the documentation to the root folder of the monorepo,
it doesn't make sense to keep building the HTML manual and the man pages
when building checkbox-ng Debian package.

Since translations have never been used in checkbox either, removing the
part that deals with them (po and pot files).