Add anchors to each fields for every Checkbox units
The fields are described as definition lists in HTML. This is, by
default, rendered without any id anchor in HTML.

Manually adding a reference in ReStructuredText allows to point a user
to the field by using the HTML id anchor. For instance:

/reference/units/test-plan/#test-plan-bootstrap-include-field

will point to the bootstrap_include field of the test plan unit.

These can also be used in rst directly by using :ref:. See [1] for more
info.

[1]
https://sublime-and-sphinx-guide.readthedocs.io/en/latest/references.html#add-links