Add link anchors to documentation and fix broken link to removed Plainbox docs (#543)
* Add anchors to each fields for every Checkbox units

The fields are described as definition lists in HTML. This is, by
default, rendered without any id anchor in HTML.

Manually adding a reference in ReStructuredText allows to point a user
to the field by using the HTML id anchor. For instance:

/reference/units/test-plan/#test-plan-bootstrap-include-field

will point to the bootstrap_include field of the test plan unit.

These can also be used in rst directly by using :ref:. See [1] for more
info.

[1]
https://sublime-and-sphinx-guide.readthedocs.io/en/latest/references.html#add-links

* Remove unused references inside flags section

Since there is now a reference to the `flags` section, these
sub-references may not be useful.

* Fix link to category-overrides reference

The plainbox documentation website has been decomissioned, so the
external link does not work anymore. Point to an internal reference
instead.

* Remove unnecessary table of contents

* Add refs for the different flags in the job unit