Run commands from the docs/ directory
Following a discussion with the documentation team that led to the
creation of a PR on their side[1], make sure to be in the docs/
directory before triggering the different commands. This should make
woke work as expected.

Please note that according to woke doc[2], the parameter to set this
is "workdir".

[1] https://github.com/canonical/sphinx-docs-starter-pack/pull/40
[2] https://github.com/marketplace/actions/run-woke