Add ability to get a numerical Intel gen number
Some feautures we test for may be heavily dependent on the CPU
generation. For instance, a media stress test for Chromium seems to
max out VECS on Gen12+ while it maxes out VCS on older generations.

Adding a command line argument to print a simple number allows tests to
be more flexible at runtime.