Change: re-structured the system confinement test
The original snappy/test-snap-confinement-mode is actually testing
the system's confinement whether it is "strict". If it is, test
passes. Otherwise, the missing feature to make system in strict
confinement will be listed.

This commit renamed the job ID and purpose to make it more clear.
Also, restructured the python script to put the original test into
SystemConfinement class as a test and added docstring to make it
easier to understand and maintain.