Rephrase Checkbox remote UI interrupt option
In its current state, the session running on the agent is paused when
the controller disconnects, and resumed when the controller reconnects
in order to perform some kind of sign off for each job run.

This may change in the future, but for the moment, renaming the UI
choice to better reflect current reality.