Install metabox provider in overlay
This installs the metabox provider in an overlay that is
automounted to the checkbox snap at boot. This is done to
avoid sideloading but also having a persistent install
of the metabox provider to be persistent across reboots.
Also, this uses the default loading mechanism of the snap
and it is closer to what it will do "in the wild"