Add Checkbox commands tutorial (#635)
* Add Checkbox commands tutorial

Fix CHECKBOX-766

* Remove "Using Checkbox" how-to

This document is not necessary anymore since we have a dedicated
tutorial that covers all of these sections now.

* Add more info about bootstrapping in glossary and link to resource unit

* Feedback from Max

* Fixup tings after removing using.rst

After removing using.rst, a reference to the `run` subcommand was broken
in the "Understanding Checkbox". The ref now points to the tutorial.

Also removing reference to `using` in the index page since it no longer
exists.