Add daily process to tag edge releases
edge_tag.py generates an edge release tag based on the latest release
tag and the number of commits since, and pushes it.

For instance, if the last release is "v2.9.1", and there have been 6 new
commits since, the script will compute tag "v2.9.2-edge+6", tag the
latest commit in HEAD with it, and push the tags.

Fix CHECKBOX-807