Refine Canary test plan (BugFix) (#716)
* Refine Canary test plan

Remove a few jobs from the Canary test plan.

Some jobs are meant to be run on desktop images and do not run well (or
at all) on Ubuntu Core:

- the graphics-related ones
- i2c/i2c-device-detect (although it could technically run on a desktop
image, it requires a connected I²C device connected to run properly)

On the other hand, some jobs are meant to run on UC and not on desktop
images:

- ubuntu_core_features
- kernel-snap/booted-kernel-matches-current-grub
- parts_meta_info_attachment

A lot of the snap-related tests make requests to the snap store, and
it's preferable to avoid this to avoid false negative in the case where
the snap store is not responding.

The watchdog tests are meant to be run only on devices where the
watchdog device is enabled and properly setup. Watchdog is not part of
the coverage for laptops and desktops.

This commit also renames the pxu file to be aligned with the convention
used for other units in this provider.

Fix CHECKBOX-837

* Remove location/status job from canary test plan

Remove location job (and its after-suspend sibling) as it relies on a
snap (locationd) that we do not support anymore and that is therefore
not installed. As a result, the jobs will always be skipped.

* Remove disk/storage_device_.* test from canary test plan

Stress test that takes a long time to run and does not bring any benefit
in terms of testing Checkbox proper.