Add requirements for the snap name and revisions to allow jobs to be skipped
This requirement trick allows to specify the stable_rev/base_rev and
original_installed_rev to check for the snap being tested.

This allows to efficiently skip the job if said snap has the same source
and destination revision.

Thanks to this, these jobs are not marked as failed anymore, but instead
skipped.