Add: snaps confinement test (new) (#663)
* Change: re-structured the system confinement test

The original snappy/test-snap-confinement-mode is actually testing
the system's confinement whether it is "strict". If it is, test
passes. Otherwise, the missing feature to make system in strict
confinement will be listed.

This commit renamed the job ID and purpose to make it more clear.
Also, restructured the python script to put the original test into
SystemConfinement class as a test and added docstring to make it
easier to understand and maintain.

* Change: replace logging string from format() to %

* Add: snaps confinement tests

Add SnapsConfinement class in snap_confinement_test.py to check
all the installed snaps' confinement, devmode, and revision.

* Add: snappy-snap-automated to desktop test plans

* Change: apply suggestions from kissiel

* Change: Flatten simple classes to test functions