fix the race condition when resuming via remote
There was a chance that when the controller reconnected to the agent,
it would get an information that something is running, so naturally it
would proceed with waiting and finishing, but this "wait and finish"
assumed the job was running, which may have finished by that point.
This could happen only on resume, cause otherwise the controller would
control the execution, so there would be no race.