Removing kernel snap refresh/revert test from automated UC test plan
This test plan is used by both QA and Cert. QA team faces issues when
running the kernel refresh/revert snap tests on early builds because the
kernel team may not have placed all the expected kernel snap updates in
the right channels (they may use private branches for the tests, or
share the same channels for different devices for hte same customer).

All of this is resolved by the time the devices are sent to Cert lab, so
the tests should be added to the SRU test plan instead.