Fix snappy-snap-automated misplaced to manual plan (BugFix) (#801)
Fix snappy-snap-automated misplaced to manual plan

snappy-snap-automated was added into client cert desktop plan with pull
request https://github.com/canonical/checkbox/pull/663. When nesting to
the desktop plan, it was misplaced to manual plan on
client-cert-desktop-22-04.pxu. 20-04 and 18-04 didn't misplace. This
commit fixes it to move the snappy-snap-automated to auto plan.