Modify the wwan related test cases.
The mobilebroadband-cert-automated test case is only
part of the wwan test plan. We hope to replace
mobilebroadband-cert-automated with wwan-automated.

The relevant test plans we currently use include:
client-cert-desktop-22-04-automated
client-cert-desktop-20-04-automated
client-cert-desktop-18-04-automated

I have replaced mobilebroadband-cert-automated in them
with wwan-automated, and added after-sudpend-wwan-
automated to ensure that wwan is tested before and after
suspend.

In addition, for mobilebroadband-cert-manual in the Manual
test plan, when the SIM is inserted into the prototype in the
desktop, the system should automatically recognize wwan and
SIM card, and we do not need to perform manual testing. So I
deleted mobilebroadband-cert-manual in the relevant test
plans. At the same time, since wwan-automated already
contains wwan-manual cases, I deleted the wwan-manual in
these test plans:
client-cert-desktop-22-04-manual
client-cert-desktop-20-04-manual
client-cert-desktop-18-04-manual