Dont require gpg pass when running tests (BugFix) (#843)
properly mock subprocess calls in test_provider_manager

Previously when the tests run, if the host had a GPG key in its databse
that was protected by a password, the operator was asked to provide the
password. This was due to subprocess.call not being mocked properly in
the provider_manager tests.

This patch adds the mocking.