Add manifest entries to test gadget/kernel/snapd snaps
The snap-refresh-*-from-[stable|base]-rev jobs are executed if the
revision of the related snap (gadget, kernel, snapd) currently installed
on the device is different from the revision in base version and/or
stable channel.

It is better to provide an option to decide whether or not these tests
should be executed, per device.

This is achieved using three manifest entries (one per snap type).

If these entries are explicitely set to True in the manifest, they will
be executed (if they match the condition described above, of course).
Otherwise, they will be skipped.